requirements.txt
pydantic>=2.0.0
PyQt6>=6.0.0
pytest>=7.0.0
pytest-qt>=4.0.0
pytest-cov>=4.0.0
python-dateutil>=2.8.2
pandas>=1.3.0
PyYAML>=6.0
